refactor: move tast classes in the proper folder
